Fix multiple releases building same cmssw-ib, etc
There is an issue when parallel building of IBs for the same architecture
happens.  This is because two different releases will still have a package
which is called cms+cmssw-ib+1.0.0, which will therefore conflict.

This fix makes sure that cmssw-ib, and all other related packages have a
different version based on the version of CMSSW they are based on.