Avoid race conditions when installing IBs.
This avoids race conditions when syncing IBs. They could have happened in case
the switch between one week and the other on the repository side was not
synchronized with the switch on the installation side. These set of changes
make sure that a new local installation is used every week, and we switch from
one week to the other only when the switch on the repository side happened.

This still lacks garbage collecting the old local areas.